students/k3340/laboratory_works/Bakirova_Shoola/LR2-3/project_library/requirements.txt
altgraph==0.17
auto-py-to-exe==2.7.8
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2018.4.16
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==2.0.6
django-allauth==0.43.0
django-cors-headers==2.2.0
django-rest-swagger==2.1.2
django-summernote==0.8.8.6
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djoser==1.1.5
Eel==0.12.4
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
idna==2.10
inflection==0.5.1
itypes==1.2.0
jieba==0.39
Jinja2==2.11.2
Library==0.0.0
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
pbr==3.1.1
pefile==2019.4.18
psycopg2==2.8.6
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.1.2
